Rename _io module to i_o and update imports
Renamed the openwave._io package to openwave.i_o throughout the codebase, including all imports, documentation, and references. Updated related documentation, dependency maps, and experiment scripts to use the new module path. Removed unused or deprecated modules under openwave/force, openwave/heat, openwave/matter, openwave/motion, and openwave/photon.
