+++ docs/Makefile
@@ -17,10 +17,10 @@ apidoc:
 	@echo "Regenerating API documentation from source..."
 	@rm -rf api/*.rst
 	@python -m sphinx.ext.apidoc -f -e -M -o api ../openwave
-	@echo "Creating openwave._io.rst..."
-	@printf "openwave.\\\\_io\n=============\n\n.. automodule:: openwave._io\n\n\n.. rubric:: Modules\n\n.. autosummary::\n   :toctree:\n   :recursive:\n\n   cli\n   render\n" > api/openwave._io.rst
-	@echo "Fixing openwave.rst to include _io module..."
-	@grep -q "_io" api/openwave.rst || perl -i.bak -pe 's/^(   common)$$/   _io\n$$1/' api/openwave.rst && rm -f api/openwave.rst.bak
+	@echo "Creating openwave.i_o.rst..."
+	@printf "openwave.\\\\i_o\n=============\n\n.. automodule:: openwave.i_o\n\n\n.. rubric:: Modules\n\n.. autosummary::\n   :toctree:\n   :recursive:\n\n   cli\n   render\n" > api/openwave.i_o.rst
+	@echo "Fixing openwave.rst to include i_o module..."
+	@grep -q "i_o" api/openwave.rst || perl -i.bak -pe 's/^(   common)$$/   i_o\n$$1/' api/openwave.rst && rm -f api/openwave.rst.bak
 	@echo "API documentation regenerated"
 
 # Generate dependency graphs
@@ -30,8 +30,8 @@ deps:
 
 # Complete documentation build (recommended)
 docs: apidoc
-	@echo "Re-adding _io to openwave.rst (in case it was removed)..."
-	@grep -q "^   _io$$" api/openwave.rst || perl -i.bak -pe 's/^(   common)$$/   _io\n$$1/' api/openwave.rst && rm -f api/openwave.rst.bak
+	@echo "Re-adding i_o to openwave.rst (in case it was removed)..."
+	@grep -q "^   i_o$$" api/openwave.rst || perl -i.bak -pe 's/^(   common)$$/   i_o\n$$1/' api/openwave.rst && rm -f api/openwave.rst.bak
 	@echo "Building complete documentation..."
 	@$(SPHINXBUILD) -M html "$(SOURCEDIR)" "$(BUILDDIR)" $(SPHINXOPTS) $(O)
 	@echo "Documentation built at: $(BUILDDIR)/html/index.html"
