--- openwave/force/__init__.py
@@ -1,10 +1,0 @@
-"""
-Force modules for OpenWave Energy Wave Theory simulations.
-
-This package contains the fundamental components:
-- electric: Electric force calculations and representations
-- magnetic: Magnetic force calculations and representations
-- gravitational: Gravitational force calculations and representations
-- strong: Strong force calculations and representations
-- orbital: Orbital force calculations and representations
-"""
