--- openwave/heat/__init__.py
@@ -1,3 +1,0 @@
-"""
-Heat module for thermal energy simulations.
-"""
