--- openwave/matter/__init__.py
@@ -1,10 +1,0 @@
-"""
-Matter modules for OpenWave Energy Wave Theory simulations.
-
-This package contains the fundamental components:
-- fundamental_particle: Fundamental particle representations and calculations
-- standalone_particle: Standalone particle representations and calculations
-- composite_particle: Composite particle representations and calculations
-- atom: Atom representations and calculations
-- molecule: Molecule representations and calculations
-"""
