--- openwave/motion/__init__.py
@@ -1,3 +1,0 @@
-"""
-Motion module for kinetic energy simulations.
-"""
