--- openwave/photon/__init__.py
@@ -1,3 +1,0 @@
-"""
-Photon module for photon energy simulations.
-"""
