+++ openwave/xperiments/_archives/spring_mass/spring_euler.py
@@ -18,7 +18,7 @@ import taichi as ti
 import time
 
 from openwave.common import colormap, constants
-from openwave._io import render
+from openwave.i_o import render
 
 import openwave.xperiments._archives.spring_mass.spacetime_medium as medium
 import openwave.xperiments._archives.spring_mass.spacetime_ewave_euler as ewave
