+++ openwave/xperiments/c_vector_field/particle.py
@@ -1,7 +1,15 @@
 """
 PARTICLE MODULE
 
-This module defines the WaveCenter class for simulating fundamental particle
+Matter modules for OpenWave Energy Wave Theory simulations.
+This module defines the WaveCenter class for simulating fundamental particle.
+
+This package contains the fundamental components:
+- fundamental_particle: Fundamental particle representations and calculations
+- standalone_particle: Standalone particle representations and calculations
+- composite_particle: Composite particle representations and calculations
+- atom: Atom representations and calculations
+- molecule: Molecule representations and calculations
 """
 
 import taichi as ti
