+++ openwave/xperiments/d_anti_gravity/_launcher.py
@@ -6,7 +6,7 @@ import sys
 
 print("Work in Progress XPERIMENT")
 print("- Gravity Attenuation by Proton Vibration")
-print("- Read this paper for more context:")
+print("- Read this research paper for more context:")
 print(
     "  https://www.researchgate.net/publication/396329338_Method_for_Controlling_the_Motion_of_Non-Magnetic_Objects_Utilizing_Proton_Vibration_Energy"
 )
